requirements.txt
Django
Jinja2
Markdown
MarkupSafe
beautifulsoup4
certifi
chardet
coreapi
coreschema
django-mathfilters
django-admin-shortcuts
django-admin-numeric-filter
django-cors-headers
django-enumfields
django-extensions-too
django-filter
django-formtools
django-rest-swagger
django-sslserver
django-widget-tweaks
djangorestframework
exitstatus
idna
itypes
openapi-codec
pip
pycryptodome
pytz
django-colorfield
requests
setuptools
simplejson
soupsieve
uritemplate
urllib3
django-heroku
Pillow
mdx_wikilink_plus
martor
django-ordered-model
django-treewidget
django-mptt
djangorestframework